Change the URL to the master file
The previous master file on the qtproject domain is now redirecting to qt.io. However QNetworkRequest won't automatically follow such a redirect (see API) and therefore the trivial solution is to replace the URL here. The better solution is to handle the redirect and/or allow overriding the URL from a setting.